added two small fixes:
one that happens with issue #62 on contracts.coffee.
In the _blame function it failes when orig.server that is passed is a string, I added an additional check that wraps the string in a ModuleName which results in correct behaviour.

another bug happens in setExported, where when orig is not set to null in each iteration of the for loop, it could set wrong information in the WeakMap if the value isn't a function or an object. (such like if it were a string), and in the case of a string it could crash, since keys in WeakMap can't be strings.